activation: Remove deprecated best provider option (#5189)
## Motivation
Closes #4801

## Changes
- remove `-1` as option for `providerID`
- remove autobenchmarking

## Test Plan
- removed tests for autobenchmarking
- updated existing tests to reject `-1` as valid option for Provider ID

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed
- [x] Update [changelog](../CHANGELOG.md) as needed

+++ activation/e2e/nipost_test.go
@@ -145,7 +145,7 @@ func TestNIPostBuilderWithClients(t *testing.T) {
 
 	opts := activation.DefaultPostSetupOpts()
 	opts.DataDir = t.TempDir()
-	opts.ProviderID.SetInt64(int64(initialization.CPUProviderID()))
+	opts.ProviderID.SetUint32(initialization.CPUProviderID())
 	opts.Scrypt.N = 2 // Speedup initialization in tests.
 	initPost(t, logger.Named("manager"), mgr, opts)
 
@@ -318,7 +318,7 @@ func TestNewNIPostBuilderNotInitialized(t *testing.T) {
 
 	opts := activation.DefaultPostSetupOpts()
 	opts.DataDir = t.TempDir()
-	opts.ProviderID.SetInt64(int64(initialization.CPUProviderID()))
+	opts.ProviderID.SetUint32(initialization.CPUProviderID())
 	opts.Scrypt.N = 2 // Speedup initialization in tests.
 	t.Cleanup(launchPostSupervisor(t, logger, grpcCfg, opts))
